tests - Refactor common.
