+++ tests/test_record.py
@@ -13,15 +13,7 @@ import requests
 import responses
 from mock import MagicMock, patch
 
-
-def _reset_environ():
-    """Reset the environment with test values."""
-    os.environ["DEN_ACCESS_TOKEN"] = "TEST"
-    os.environ["DEN_FORECAST_IO_API_KEY"] = "TEST"
-    os.environ["DEN_LAT"] = "39.952447"
-    os.environ["DEN_LON"] = "-75.1635083"
-
-
+from . import _reset_environ
 _reset_environ()
 from den import record, weather
